handle end-point forward and backward navigation
This commit keeps the forward and backward point history from piling up with
garbage when the user tries to navigate back from "/" or forward on a line with
no field on it, in which case the field and the root are identical.

Also add some comments in `hdf5-display-fields' to delineate sections of the
function.