Handle no-group no-dataset edge case
see
https://docs.hdfgroup.org/archive/support/ftp/HDF5/examples/files/exbyapi/index.html

I can't remember which one now, but there is a file (or maybe several) in this
set of test HDF5 files that needs the code to handle `obj.shape is None`
gracefully.

We assume in the python code that if a field is not a group, then it is a
dataset, but that may not be strictly true.